separated examples code completely
