+++ example/port_scan.dart
@@ -1,6 +1,6 @@
 import 'package:logging/logging.dart';
-import '../lib/src/network_tools_utils.dart';
 import 'package:network_tools/network_tools.dart';
+import '../example/lib/example_utils.dart';
 
 Future<void> main() async {
   enableExampleLogging();
